refactor(types): relocate most of `ballot_card.rs` back in `ballot-interpreter`
It's not used by other things and is not terribly likely to be.
